refactor userUpdate functions
