+++ functions/src/models.ts
@@ -1,3 +1,5 @@
+import * as functions from 'firebase-functions'
+
 // Models can be imported from the main package for use here
 // NOTE 1 - this requires adjustment main src in package.json
 // NOTE 2 - shorthand @OAModels notation defined in tsconfig
@@ -5,6 +7,8 @@ export * from '../../src/models/common.models'
 export * from '../../src/models/user.models'
 export * from '../../src/models/howto.models'
 export * from '../../src/models/events.models'
+
+export type IDBDocChange = functions.Change<FirebaseFirestore.DocumentSnapshot>
 
 // TODO - handle import from src/models (currently breaks ts setup)
 // but possibly fixed by backend improvements pr
